refactor: rename series into tv_shows and adjust some folder names
