+++ .github/workflows/cd-web.yml
@@ -29,7 +29,9 @@ jobs:
           dart run build_runner build --delete-conflicting-outputs
 
       - name: Flutter build web
-        run: flutter build web --web-renderer canvaskit --release --base-href "${{ secrets.BASE_HREF }}"
+        run:
+          flutter build web --web-renderer canvaskit --release --base-href
+          "/screen_pal/"
 
       - name: Upload build
         uses: actions/upload-artifact@v2
