+++ lib/interfaces/providers/movies/movie_list_providers.dart
@@ -1,10 +1,10 @@
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 import 'package:screen_pal/core/entities/movie.dart';
-import 'package:screen_pal/core/use_cases/movie/get_now_playing_movies.dart';
-import 'package:screen_pal/core/use_cases/movie/get_popular_movies.dart';
-import 'package:screen_pal/core/use_cases/movie/get_top_rated_movies.dart';
-import 'package:screen_pal/core/use_cases/movie/get_upcoming_movies.dart';
+import 'package:screen_pal/core/use_cases/movies/get_now_playing_movies.dart';
+import 'package:screen_pal/core/use_cases/movies/get_popular_movies.dart';
+import 'package:screen_pal/core/use_cases/movies/get_top_rated_movies.dart';
+import 'package:screen_pal/core/use_cases/movies/get_upcoming_movies.dart';
 import 'package:screen_pal/service_locator/locator.dart';
 
 part 'movie_list_providers.g.dart';
