+++ lib/service_locator/registry/registry.dart
@@ -5,13 +5,13 @@ import 'package:shared_preferences/shared_preferences.dart';
 import 'package:screen_pal/core/repo/genres_repo.dart';
 import 'package:screen_pal/core/repo/movies_repo.dart';
 
-import 'package:screen_pal/core/use_cases/genre/get_movie_genres.dart';
-import 'package:screen_pal/core/use_cases/genre/get_tv_show_genres.dart';
-import 'package:screen_pal/core/use_cases/movie/get_movie_detail.dart';
-import 'package:screen_pal/core/use_cases/movie/get_now_playing_movies.dart';
-import 'package:screen_pal/core/use_cases/movie/get_popular_movies.dart';
-import 'package:screen_pal/core/use_cases/movie/get_top_rated_movies.dart';
-import 'package:screen_pal/core/use_cases/movie/get_upcoming_movies.dart';
+import 'package:screen_pal/core/use_cases/genres/get_movie_genres.dart';
+import 'package:screen_pal/core/use_cases/genres/get_tv_show_genres.dart';
+import 'package:screen_pal/core/use_cases/movies/get_movie_detail.dart';
+import 'package:screen_pal/core/use_cases/movies/get_now_playing_movies.dart';
+import 'package:screen_pal/core/use_cases/movies/get_popular_movies.dart';
+import 'package:screen_pal/core/use_cases/movies/get_top_rated_movies.dart';
+import 'package:screen_pal/core/use_cases/movies/get_upcoming_movies.dart';
 
 import 'package:screen_pal/infrastructures/api/tmdb_dio.dart';
 import 'package:screen_pal/infrastructures/cache/theme_mode_cache.dart';
