+++ test/core/use_cases/movies/get_now_playing_movies_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 
-import 'package:screen_pal/core/use_cases/movie/get_now_playing_movies.dart';
+import 'package:screen_pal/core/use_cases/movies/get_now_playing_movies.dart';
 
 import '../../../helpers/mocks/repo/repositories.mocks.dart';
